[update]rename runtime-compiler
