+++ packages/webpack-plugin/lib/template-compiler/index.js
@@ -1,5 +1,4 @@
 const compiler = require('./compiler')
-const runtimeCompiler = require('./runtime-compiler')
 const loaderUtils = require('loader-utils')
 const bindThis = require('./bind-this').transform
 const InjectDependency = require('../dependency/InjectDependency')
@@ -48,9 +47,7 @@ module.exports = function (raw) {
     )
   }
 
-  const _compiler = options.runtimeCompile ? runtimeCompiler : compiler
-
-  const parsed = runtimeCompiler.parse(raw, {
+  const parsed = compiler.parse(raw, {
     warn,
     error,
     componentsAbsolutePath: options.componentsAbsolutePath || {},
@@ -84,7 +81,7 @@ module.exports = function (raw) {
     }
   }
 
-  let result = runtimeCompiler.serialize(ast, meta)
+  let result = compiler.serialize(ast, meta)
 
   if (options.isNative || mpx.forceDisableInject) {
     return result
@@ -103,7 +100,7 @@ module.exports = function (raw) {
             if (!injectSlots[key][slotTarget]) {
               injectSlots[key][slotTarget] = []
             }
-            const rawRenderFn = runtimeCompiler.genElement(node, options.usingComponents)
+            const rawRenderFn = compiler.genElement(node, options.usingComponents)
             const code = bindThis(rawRenderFn, {
               needCollect: true,
               ignoreMap: meta.wxsModuleMap
@@ -122,7 +119,7 @@ module.exports = function (raw) {
     // 运行时编译组件返回内容，render 函数内部执行会生成 virtual dom tree
     try {
       let _res
-      let renderFn = _compiler.genElement(ast)
+      let renderFn = compiler.genElement(ast)
       // console.log('the renderFn is:', renderFn)
       _res = bindThis(renderFn, {
         needCollect: true,
@@ -142,7 +139,7 @@ module.exports = function (raw) {
     const rawCode = `global.currentInject = {
       moduleId: ${JSON.stringify(options.moduleId)},
       render: function () {
-        ${_compiler.genNode(ast)}this._r();
+        ${compiler.genNode(ast)}this._r();
       }
     };\n`
     try {
@@ -161,39 +158,6 @@ module.exports = function (raw) {
       return result
     }
   }
-
-  // ${_compiler.genNode(ast)}this._r();
-//   const rawCode = `global.currentInject = {
-//     moduleId: ${JSON.stringify(options.moduleId)},
-//     render: function () {
-//       ${_compiler.genElement(ast)}
-//     }
-// };\n`
-
-//   let renderResult = {}
-
-//   try {
-//     renderResult.code = `global.currentInject = {
-//     moduleId: ${JSON.stringify(options.moduleId)},
-//     render: function () {
-//       ${_res.code}
-//     }
-// };\n`
-//     renderResult.propKeys = _res.propKeys
-//     // renderResult = bindThis(rawCode, {
-//     //   needCollect: true,
-//     //   ignoreMap: meta.wxsModuleMap
-//     // })
-//   } catch (e) {
-//     error(`Invalid render function generated by the template, please check!\n
-// Template result:
-// ${result}\n
-// Error code:
-// ${rawCode}
-// Error Detail:
-// ${e.stack}`)
-//     return result
-//   }
 
   // todo 此处在loader中往其他模块addDep更加危险，考虑修改为通过抽取后的空模块的module.exports来传递信息
   let globalInjectCode = renderResult.code + '\n'
